Skip & show notification when a song fails to load
When a song fails to load (e.g., a 4xx or 5xx HTTP status), show a
user-friendly notification.

If the play mode is not "one song only", skip to the next track. If there is a
string of failures, stop trying to skip and show a different notification
(much like the GMusic UI).

Fixes GH malept/gmusicprocurator#8.